# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Enterprise Knowledge Intelligence Platform
  version: 1.0.0
paths:
...
  /api/v1/auth/register:
    post:
      tags:
      - authentication
      summary: Register
      operationId: register_api_v1_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/login:
    post:
      tags:
      - authentication
      summary: Login
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/me:
    get:
      tags:
      - authentication
      summary: Me
      operationId: me_api_v1_auth_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/documents:
    get:
      tags:
      - documents
      summary: Documents
      operationId: documents_api_v1_documents_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: X-Workspace-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Workspace-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentRead'
                title: Response Documents Api V1 Documents Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - documents
      summary: Upload Document
      operationId: upload_document_api_v1_documents_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: X-Workspace-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Workspace-Id
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_document_api_v1_documents_post'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/documents/{document_id}:
    get:
      tags:
      - documents
      summary: Document Detail
      operationId: document_detail_api_v1_documents__document_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Document Id
      - name: X-Workspace-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Workspace-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - documents
      summary: Delete Document
      operationId: delete_document_api_v1_documents__document_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Document Id
      - name: X-Workspace-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Workspace-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/search:
    post:
      tags:
      - search
      summary: Search
      operationId: search_api_v1_search_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: X-Workspace-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Workspace-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/research:
    post:
      tags:
      - research
      summary: Create
      operationId: create_api_v1_research_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: X-Workspace-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Workspace-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResearchRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - research
      summary: List Research
      operationId: list_research_api_v1_research_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: X-Workspace-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Workspace-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResearchRead'
                title: Response List Research Api V1 Research Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/research/{job_id}:
    get:
      tags:
      - research
      summary: Read
      operationId: read_api_v1_research__job_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Job Id
      - name: X-Workspace-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Workspace-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/evaluation:
    post:
      tags:
      - evaluation
      summary: Create
      operationId: create_api_v1_evaluation_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: X-Workspace-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Workspace-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluationRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - evaluation
      summary: List Runs
      operationId: list_runs_api_v1_evaluation_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: X-Workspace-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Workspace-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EvaluationRead'
                title: Response List Runs Api V1 Evaluation Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/jobs:
    get:
      tags:
      - jobs
      summary: Jobs
      operationId: jobs_api_v1_jobs_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: X-Workspace-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Workspace-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobRead'
                title: Response Jobs Api V1 Jobs Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/jobs/{job_id}:
    get:
      tags:
      - jobs
      summary: Job
      operationId: job_api_v1_jobs__job_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Job Id
      - name: X-Workspace-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Workspace-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    Body_upload_document_api_v1_documents_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      type: object
      required:
      - file
      title: Body_upload_document_api_v1_documents_post
    DocumentRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        workspace_id:
          type: string
          format: uuid
          title: Workspace Id
        title:
          type: string
          title: Title
        status:
          type: string
          title: Status
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        created_by:
          type: string
          format: uuid
          title: Created By
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - workspace_id
      - title
      - status
      - description
      - created_by
      - created_at
      - updated_at
      title: DocumentRead
    DocumentVersionRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        document_id:
          type: string
          format: uuid
          title: Document Id
        version_number:
          type: integer
          title: Version Number
        filename:
          type: string
          title: Filename
        mime_type:
          type: string
          title: Mime Type
        size_bytes:
          type: integer
          title: Size Bytes
        checksum_sha256:
          type: string
          title: Checksum Sha256
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - document_id
      - version_number
      - filename
      - mime_type
      - size_bytes
      - checksum_sha256
      - created_at
      title: DocumentVersionRead
    EvaluationCase:
      properties:
        question:
          type: string
          minLength: 2
          title: Question
        expected_answer:
          type: string
          minLength: 1
          title: Expected Answer
      type: object
      required:
      - question
      - expected_answer
      title: EvaluationCase
    EvaluationRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        workspace_id:
          type: string
          format: uuid
          title: Workspace Id
        user_id:
          type: string
          format: uuid
          title: User Id
        name:
          type: string
          title: Name
        status:
          type: string
          title: Status
        metrics_json:
          additionalProperties: true
          type: object
          title: Metrics Json
        config_json:
          additionalProperties: true
          type: object
          title: Config Json
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - workspace_id
      - user_id
      - name
      - status
      - metrics_json
      - config_json
      - created_at
      - updated_at
      title: EvaluationRead
    EvaluationRequest:
      properties:
        name:
          type: string
          maxLength: 160
          minLength: 2
          title: Name
        cases:
          items:
            $ref: '#/components/schemas/EvaluationCase'
          type: array
          maxItems: 100
          minItems: 1
          title: Cases
      type: object
      required:
      - name
      - cases
      title: EvaluationRequest
    EvidenceItem:
      properties:
        chunk_id:
          type: string
          format: uuid
          title: Chunk Id
        document_id:
          type: string
          format: uuid
          title: Document Id
        document_title:
          type: string
          title: Document Title
        content:
          type: string
          title: Content
        score:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Score
        metadata:
          additionalProperties: true
          type: object
          title: Metadata
      type: object
      required:
      - chunk_id
      - document_id
      - document_title
      - content
      - score
      title: EvidenceItem
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    JobRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        workspace_id:
          type: string
          format: uuid
          title: Workspace Id
        document_version_id:
          type: string
          format: uuid
          title: Document Version Id
        status:
          type: string
          title: Status
        stage:
          type: string
          title: Stage
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
        result_json:
          additionalProperties: true
          type: object
          title: Result Json
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - workspace_id
      - document_version_id
      - status
      - stage
      - error_message
      - result_json
      - created_at
      - updated_at
      title: JobRead
    MessageResponse:
      properties:
        message:
          type: string
          title: Message
      type: object
      required:
      - message
      title: MessageResponse
    ResearchRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        workspace_id:
          type: string
          format: uuid
          title: Workspace Id
        user_id:
          type: string
          format: uuid
          title: User Id
        question:
          type: string
          title: Question
        status:
          type: string
          title: Status
        report_markdown:
          anyOf:
          - type: string
          - type: 'null'
          title: Report Markdown
        result_json:
          additionalProperties: true
          type: object
          title: Result Json
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - workspace_id
      - user_id
      - question
      - status
      - report_markdown
      - result_json
      - error_message
      - created_at
      - updated_at
      title: ResearchRead
    ResearchRequest:
      properties:
        question:
          type: string
          maxLength: 4000
          minLength: 5
          title: Question
      type: object
      required:
      - question
      title: ResearchRequest
    SearchRequest:
      properties:
        query:
          type: string
          maxLength: 4000
          minLength: 2
          title: Query
        top_k:
          anyOf:
          - type: integer
            maximum: 50.0
            minimum: 1.0
          - type: 'null'
          title: Top K
        document_ids:
          anyOf:
          - items:
              type: string
              format: uuid
            type: array
          - type: 'null'
          title: Document Ids
      type: object
      required:
      - query
      title: SearchRequest
    SearchResponse:
      properties:
        answer:
          type: string
          title: Answer
        evidence:
          items:
            $ref: '#/components/schemas/EvidenceItem'
          type: array
          title: Evidence
        sufficient_evidence:
          type: boolean
          title: Sufficient Evidence
        abstained:
          type: boolean
          title: Abstained
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
      type: object
      required:
      - answer
      - evidence
      - sufficient_evidence
      - abstained
      title: SearchResponse
    TokenResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
          default: bearer
        expires_in:
          type: integer
          title: Expires In
        user:
          $ref: '#/components/schemas/UserRead'
        workspace_id:
          type: string
          format: uuid
          title: Workspace Id
      type: object
      required:
      - access_token
      - expires_in
      - user
      - workspace_id
      title: TokenResponse
    UploadResponse:
      properties:
        document:
          $ref: '#/components/schemas/DocumentRead'
        version:
          $ref: '#/components/schemas/DocumentVersionRead'
        job_id:
          type: string
          format: uuid
          title: Job Id
        status:
          type: string
          title: Status
      type: object
      required:
      - document
      - version
      - job_id
      - status
      title: UploadResponse
    UserCreate:
      properties:
        email:
          type: string
          format: email
          title: Email
        full_name:
          type: string
          maxLength: 160
          minLength: 2
          title: Full Name
        password:
          type: string
          maxLength: 128
          minLength: 12
          title: Password
        organization_name:
          type: string
          maxLength: 160
          minLength: 2
          title: Organization Name
          default: My Organization
        workspace_name:
          type: string
          maxLength: 160
          minLength: 2
          title: Workspace Name
          default: General
      type: object
      required:
      - email
      - full_name
      - password
      title: UserCreate
    UserLogin:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          title: Password
      type: object
      required:
      - email
      - password
      title: UserLogin
    UserRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        email:
          type: string
          format: email
          title: Email
        full_name:
          type: string
          title: Full Name
        is_active:
          type: boolean
          title: Is Active
        is_superuser:
          type: boolean
          title: Is Superuser
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - email
      - full_name
      - is_active
      - is_superuser
      - created_at
      title: UserRead
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login